Coonsolidate the UI in the Manufacture panel area
This is a pre-work for #1471 to reduce the impact of changes. It also aligns to creating a single processing queue.

Changes:
- Create a new ProcessListPanel that display manufacture & salvage processes
- Move the 2 child classes, ManufacturePanel & SalvagePanel, to the util package
- Building manufacture & Settlement manufacture tab panels use the new ProcessListPanel and remvoe a lot of duplicate code
- Drop the ability to create new processes directly in a Building; this is not needed when a central queue
- Only calculate potential processes once a Building has been selected
- Correct problem with Settlement ManufacturePanel not enabling the start new process button

Part of #1410
