+++ mars-sim-core/src/main/resources/messages.properties
@@ -1795,16 +1795,16 @@ TabPanelMaintenance.title							= Maintenance
 
 
 TabPanelManufacture.button.createNewProcess        = Start This Process
-TabPanelManufacture.checkbox.overrideManufacturing	= Override Manufacturing
-TabPanelManufacture.checkbox.overrideSalvaging		= Override Salvaging
+TabPanelManufacture.checkbox.overrideManufacturing	= Manual Manufacturing
+TabPanelManufacture.checkbox.overrideSalvaging		= Manual Salvaging
 TabPanelManufacture.cutOff                         = ...
 TabPanelManufacture.label                          = Manufacturing
 TabPanelManufacture.log.newProcessButton           = New Process Button
 TabPanelManufacture.title                          = Manu
 TabPanelManufacture.tooltip                        = Manufacturing
 TabPanelManufacture.tooltip.createNewProcess       = Create a New Manufacturing or salvage process
-TabPanelManufacture.tooltip.overrideManufacturing	= Prevents Settlers from Starting New Manufacturing Processes
-TabPanelManufacture.tooltip.overrideSalvaging		= Prevents Settlers from Starting New Salvage Processes
+TabPanelManufacture.tooltip.overrideManufacturing  = Prevents automatic starting of new Manufacturing Processes
+TabPanelManufacture.tooltip.overrideSalvaging	   = Prevents automatic starting of new Salvage Processes
 TabPanelManufacture.tooltip.selectAvailableProcess = Select an Available Manufacturing Process
 TabPanelManufacture.tooltip.selectBuilding         = Select a Manufacturing Building
